requirements.txt
psycopg2-binary==2.9.9
cx_Oracle==8.3.0
pandas==1.5.3
PyYAML==6.0
awswrangler==2.17.0
